vectorClass.js
This is a vector class in javascript it allows the user to create a vector of two points, scale the vector accordingly(multiply or divide), it allows for dot multiplication, adding and subtracting two vectors together and the magnitude between two vectors(distance). Aside from magnitude and dot which result in a single value result the other functions return a new vector with the appropriate point values.