collisionForces.js
This describes the interaction between two circles of equal mass colliding. It makes use of the vector class in order to function properly. The minimum translation distance(mtd), which is the distance need to push them apart after collision(0.5), is the mass of that circle divided by the total mass of both circles. Since they are equal in this case the mtd is only 0.5. Assigns the appropriate velocity to each circle based of the equations here https://en.wikipedia.org/wiki/Elastic_collision#Two-dimensional